Split ScanDetail and Settings pages into folder modules
Convert the two largest pages into colocated folder modules; the
preact-iso lazy imports resolve ./ScanDetail and ./Settings to the
folder index.tsx, so route wiring is unchanged:

- ScanDetail/: DecisionDialog, ReleaseRecommendation, DiffWorkbench,
  FindingsSection, ReportSections, types + index orchestrator.
- Settings/: GithubAppSection, ReleaseTargetForm, NpmConnectionSection
  + index orchestrator.

Promote the local styled select to a shared src/components/Select.tsx
primitive. Severity/format utils import from src/lib. Pure refactor,
no behavior change.
